+++ crates/cow-shed/src/eip712/mod.rs
@@ -4,6 +4,6 @@ mod domain;
 mod hash;
 pub mod sol_types;
 
-pub use domain::cow_shed_domain_separator;
-pub use hash::{execute_hooks_message_hash, hash_to_sign};
+pub use domain::{cow_shed_domain_separator, cow_shed_eip712_domain};
+pub use hash::execute_hooks_signing_hash;
 pub use sol_types::{Call as SolCall, ExecuteHooks};